Current thoughts on the multi go routine-safe version
I think the base package is good: you can step timings, not the profile.
This means nothing has to keep track of current - it's up to the user, and
not a large burden. I considered keeping the old API as well, but I think
this is a better way overall.

The main difficulty is the GAE interface. I'm not sure how to associate the
contexts with Timings. One way is the one in this commit: move contexts to
be with Timings. This presents the problem that the _gae.Profile is no
longer a context. Also, there's no obvious way to tie goon calls to the
correct Timing.